+++ proj_LittleGameplayTweaks/code/Config.cs
@@ -421,8 +421,8 @@ namespace LittleGameplayTweaks
             );
             InteractableRedSoupAmount = ConfigFileLGT.Bind(
                 "Interactables",
-                "RedToWhite Cauldron extra item amount",
-                1,
+                "RedToWhite Cauldron Extra item amount",
+                0,
                 "This is in addition to the 3 that Vanilla pays out with\nie If you want 5 items total, set it to 2.  Set to 0 to disable."
             );
             InteractableNoLunarCost = ConfigFileLGT.Bind(